requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv
gradio
openai>=1.0.0
numpy>=1.24.0
requests>=2.31.0
azure-cosmos
azure-identity
gunicorn
